refactor: clean up header
